+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - 添加配置选项 `config.query.index` 用于控制是否显示查询结果序号 | Add configuration option `config.query.index` to control whether to display the query result ordinal.
 - 使用字段别名前缀 `__hidden__` 在查询结果中隐藏该字段 | Use the field alias prefix `__hidden__` to hide the field in the query result.
 - 修复 `.all-contributorsrc` 文件配置问题 | Fix the configuration problem of the `.all-contributorsrc` file.
+- 添加多列显示块属性功能 | Add multiple column display block attribute function.
 
 ## v0.2.4/2022-03-23